rearrange exceptions
